backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pydantic
sentence-transformers
pypdf
python-multipart
